user validation

+++ taskAPI/application/controllers/task.php
@@ -33,11 +33,30 @@ class task extends CI_Controller {
 
     }
 
+    public function check_access() {
+
+        $this->load->model('sys_users_model');
+
+        $ret_msg = $this->sys_users_model->validate_access();
+
+        if($ret_msg['result'] == 'OK'){
+            return true;
+        } else {
+            return false;
+        }
+
+    }
+
     public function get($taskId=0){
 
+        $this->load->model('sys_users_model');
         $this->load->model('task_model');
 
-        $data['response'] = json_encode($this->task_model->getTasks($taskId));
+        if($this->check_access()){
+            $data['response'] = json_encode($this->task_model->getTasks($taskId), JSON_NUMERIC_CHECK);
+        } else {
+            $data['response'] = json_encode(array('result' => 'NOK_TOKEN_ERROR'));
+        }
 
         $this->load->view('task', $data);
 
@@ -45,9 +64,14 @@ class task extends CI_Controller {
 
     public function add(){
 
+        $this->load->model('sys_users_model');
         $this->load->model('task_model');
         
-        $data['response'] = $this->task_model->add();
+        if($this->check_access()){
+            $data['response'] = $this->task_model->add();
+        } else {
+            $data['response'] = json_encode(array('result' => 'NOK_TOKEN_ERROR'));
+        } 
 
         $this->load->view('task', $data);
 
@@ -55,9 +79,14 @@ class task extends CI_Controller {
 
     public function delete(){
 
+        $this->load->model('sys_users_model');
         $this->load->model('task_model');
 
-        $data['response'] = $this->task_model->delete();
+        if($this->check_access()){
+            $data['response'] = $this->task_model->delete();
+        } else {
+            $data['response'] = json_encode(array('result' => 'NOK_TOKEN_ERROR'));
+        }
 
         $this->load->view('task', $data);
 
@@ -65,13 +94,17 @@ class task extends CI_Controller {
 
     public function update(){
 
+        $this->load->model('sys_users_model');
         $this->load->model('task_model');
 
-        $data['response'] = $this->task_model->update();
+        if($this->check_access()){
+            $data['response'] = $this->task_model->update();
+        } else {
+            $data['response'] = json_encode(array('result' => 'NOK_TOKEN_ERROR'));
+        }
 
         $this->load->view('task', $data);
         
-
     }
